refactor(benchmark): move VQA generation package
