+++ dimos/benchmark/vqa/generation/pipeline.py
@@ -12,22 +12,19 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Single-frame VQA ground-truth generation and image-only evaluation."""
+"""Single-frame VQA private ground-truth generation."""
 
 from __future__ import annotations
 
-from dimos.perception.detection.type.detection2d.seg import Detection2DSeg
-from dimos.perception.vqa.evaluate import evaluate_examples
-from dimos.perception.vqa.grounding import ground_segmented_objects
-from dimos.perception.vqa.models import (
+from dimos.benchmark.vqa.generation.grounding import ground_segmented_objects
+from dimos.benchmark.vqa.generation.questions import generate_questions
+from dimos.benchmark.vqa.models import (
     CalibratedFrame,
     ObjectDetector,
     ObjectSegmenter,
-    VisualQuestionAnswerer,
-    VqaEvaluation,
     VqaExample,
 )
-from dimos.perception.vqa.questions import generate_questions
+from dimos.perception.detection.type.detection2d.seg import Detection2DSeg
 
 
 def generate_ground_truth(
@@ -44,12 +41,3 @@ def generate_ground_truth(
         segmented.extend(detection for detection in result if isinstance(detection, Detection2DSeg))
     objects = ground_segmented_objects(frame, segmented)
     return generate_questions(frame.id, objects, queries)
-
-
-def evaluate_ground_truth(
-    frame: CalibratedFrame,
-    examples: list[VqaExample],
-    answerer: VisualQuestionAnswerer,
-) -> list[VqaEvaluation]:
-    """Evaluate an answerer without passing it geometry or expected answers."""
-    return evaluate_examples(frame.image, examples, answerer)
